abstraida interfaz de usuario a una general.
