+++ src/app/Http/Controllers/Aux/AbstractAuxController.php
@@ -1,0 +1,17 @@
+<?php namespace PCI\Http\Controllers\Aux;
+
+use PCI\Http\Controllers\Controller;
+
+abstract class AbstractAuxController extends Controller
+{
+
+    /**
+     * Create a new controller instance.
+     *
+     * @method __construct
+     */
+    public function __construct()
+    {
+        $this->middleware('admin');
+    }
+}
